# api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backup Manager /api/v1",
    "version": "1.0.0",
    "summary": "The authoritative definition of the versioned /api/v1 boundary.",
    "description": "Issue #166 (B6.2). This document, not the Go handlers and not the TypeScript client, is the definition of /api/v1. Both sides are generated from it (scripts/api/generate.sh) and both are held to it by a drift check (scripts/api/check-contract-drift.sh plus the conformance tests named in docs/api/contract.md). Compatibility and deprecation rules are in docs/api/contract.md."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Every path in this document is relative to the versioned base path. /health/live and /health/ready are deliberately NOT in this contract: they sit outside /api/v1, carry no authentication and exist for an orchestrator, not for a client of this API."
    }
  ],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bm_session",
        "description": "The authenticated session. For the generic runtime profile this is apps/common/auth/local's own HTTP-only cookie; a runtime profile with a native gateway (native_auth) establishes the same session from provider identity headers accepted only from a trusted peer. Which mechanism established it is a profile concern; that a caller of a secured operation has one is not."
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "The double-submit CSRF token (apps/common/csrf). Required on every state-changing operation, and on the two read-only-in-effect operations that still open an outbound connection to a caller-supplied host."
      },
      "bootstrapToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Bootstrap-Token",
        "description": "The single-use enrollment secret printed to the runtime's own log. Only /auth/enroll accepts it."
      }
    },
...
    "schemas": {
      "ApiErrorCode": {
        "type": "string",
        "description": "Every error code this API contract knows. `wire` codes are the machine-readable tokens a server may put on the wire; `ui` codes are the shared web UI's own presentation vocabulary, which no endpoint ever emits and which is listed here so there is exactly one registry rather than a second hand-maintained copy in ui/shared. `unknown` is the sentinel an unrecognised code off the network degrades to.",
        "enum": [
          "authentication-failed",
          "ssh-host-key-changed",
          "permission-denied",
          "remote-path-missing",
          "checksum-mismatch",
          "backup-stale",
          "storage-critical",
          "version-mismatch",
          "operation-conflict",
          "unknown",
          "UNAUTHENTICATED",
          "RATE_LIMITED",
          "INVALID_REQUEST",
          "ENROLLMENT_CLOSED",
          "BOOTSTRAP_TOKEN_INVALID",
          "INTERNAL_ERROR",
          "CSRF_TOKEN_MISSING",
          "CSRF_TOKEN_MISMATCH",
          "RETENTION_PLAN_STALE",
          "RETENTION_PLAN_NOT_FOUND",
          "RETENTION_APPLY_BUSY",
          "BACKUP_SET_NOT_FOUND",
          "OPERATION_NOT_FOUND",
          "OPERATION_ALREADY_RUNNING",
          "IDEMPOTENCY_KEY_CONFLICT",
          "CONFIG_REVISION_STALE",
          "SSH_KEY_NOT_FOUND",
          "HOST_KEY_PROBE_FAILED",
          "DESTRUCTIVE_OPERATIONS_DISABLED",
          "INTERNAL",
          "ARTIFACT_NOT_FOUND",
          "ARTIFACT_NOT_QUARANTINED",
          "ARTIFACT_IRRECOVERABLE"
        ],
        "x-code-origins": {
          "authentication-failed": "ui",
          "ssh-host-key-changed": "ui",
          "permission-denied": "ui",
          "remote-path-missing": "ui",
          "checksum-mismatch": "ui",
          "backup-stale": "ui",
          "storage-critical": "ui",
          "version-mismatch": "ui",
          "operation-conflict": "ui",
          "unknown": "ui",
          "UNAUTHENTICATED": "wire",
          "RATE_LIMITED": "wire",
          "INVALID_REQUEST": "wire",
          "ENROLLMENT_CLOSED": "wire",
          "BOOTSTRAP_TOKEN_INVALID": "wire",
          "INTERNAL_ERROR": "wire",
          "CSRF_TOKEN_MISSING": "wire",
          "CSRF_TOKEN_MISMATCH": "wire",
          "RETENTION_PLAN_STALE": "wire",
          "RETENTION_PLAN_NOT_FOUND": "wire",
          "RETENTION_APPLY_BUSY": "wire",
          "BACKUP_SET_NOT_FOUND": "wire",
          "OPERATION_NOT_FOUND": "wire",
          "OPERATION_ALREADY_RUNNING": "wire",
          "IDEMPOTENCY_KEY_CONFLICT": "wire",
          "CONFIG_REVISION_STALE": "wire",
          "SSH_KEY_NOT_FOUND": "wire",
          "HOST_KEY_PROBE_FAILED": "wire",
          "DESTRUCTIVE_OPERATIONS_DISABLED": "wire",
          "INTERNAL": "wire",
          "ARTIFACT_NOT_FOUND": "wire",
          "ARTIFACT_NOT_QUARANTINED": "wire",
          "ARTIFACT_IRRECOVERABLE": "wire"
        },
        "x-error-classes": {
          "authentication": [
            "UNAUTHENTICATED",
            "BOOTSTRAP_TOKEN_INVALID"
          ],
          "authorization": [
            "ENROLLMENT_CLOSED",
            "DESTRUCTIVE_OPERATIONS_DISABLED",
            "CSRF_TOKEN_MISSING",
            "CSRF_TOKEN_MISMATCH"
          ],
          "validation": [
            "INVALID_REQUEST",
            "SSH_KEY_NOT_FOUND",
            "HOST_KEY_PROBE_FAILED"
          ],
          "conflict": [
            "RETENTION_PLAN_STALE",
            "RETENTION_APPLY_BUSY",
            "OPERATION_ALREADY_RUNNING",
            "IDEMPOTENCY_KEY_CONFLICT",
            "CONFIG_REVISION_STALE",
            "ARTIFACT_NOT_QUARANTINED",
            "ARTIFACT_IRRECOVERABLE"
          ],
          "not-found": [
            "BACKUP_SET_NOT_FOUND",
            "OPERATION_NOT_FOUND",
            "RETENTION_PLAN_NOT_FOUND",
            "ARTIFACT_NOT_FOUND"
          ],
          "throttling": [
            "RATE_LIMITED"
          ],
          "internal": [
            "INTERNAL",
            "INTERNAL_ERROR"
          ]
        }
      },
...
      "ImportSSHKeyRequest": {
        "type": "object",
        "description": "POST /ssh-keys. Sent once; the caller discards its own copy immediately.",
        "additionalProperties": false,
        "properties": {
          "private_key_pem": {
            "type": "string",
            "description": "Private key material. Write-only: it is never echoed back by any operation in this contract.",
            "writeOnly": true
          }
        },
        "required": [
          "private_key_pem"
        ]
      },
      "ImportSSHKeyResponse": {
        "type": "object",
        "description": "The reference a later create-backup-set call carries instead of key material.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "fingerprint": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "algorithm",
          "fingerprint"
        ]
      },
      "HostKeyProbeRequest": {
        "type": "object",
        "description": "POST /ssh/host-key-probe. Opens a real outbound connection, trusts nothing.",
        "additionalProperties": false,
        "properties": {
          "host": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          }
        },
        "required": [
          "host",
          "port"
        ]
      },
      "HostKeyProbeResponse": {
        "type": "object",
        "description": "What the server actually saw. Presenting it to an operator for approval is the caller's job.",
        "additionalProperties": false,
        "properties": {
          "algorithm": {
            "type": "string"
          },
          "fingerprint": {
            "type": "string"
          },
          "known_hosts_line": {
            "type": "string"
          }
        },
        "required": [
          "algorithm",
          "fingerprint",
          "known_hosts_line"
        ]
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The nested error body every operation outside /auth returns. code is stable and machine-readable; message is human-readable and MAY change without notice.",
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ApiErrorCode"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "description": "The error envelope every operation outside /auth returns. The correlation id travels in the X-Correlation-Id response header.",
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/ssh-keys": {
      "post": {
        "operationId": "importSSHKey",
        "summary": "Import a private key and get back a reference to it.",
        "tags": [
          "ssh"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSSHKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportSSHKeyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Refused: INVALID_REQUEST.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INVALID_REQUEST"
            ]
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": true,
        "x-authenticated": true
      }
    },
    "/ssh/host-key-probe": {
      "post": {
        "operationId": "probeHostKey",
        "summary": "Fetch a host key fingerprint, trusting nothing. Carries CSRF because it opens an outbound connection to a caller-supplied host.",
        "tags": [
          "ssh"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostKeyProbeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostKeyProbeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Refused: INVALID_REQUEST, HOST_KEY_PROBE_FAILED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INVALID_REQUEST",
              "HOST_KEY_PROBE_FAILED"
            ]
          },
          "401": {
            "description": "Refused: UNAUTHENTICATED.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "UNAUTHENTICATED"
            ]
          },
          "403": {
            "description": "Refused: CSRF_TOKEN_MISSING, CSRF_TOKEN_MISMATCH.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "CSRF_TOKEN_MISSING",
              "CSRF_TOKEN_MISMATCH"
            ]
          },
          "500": {
            "description": "Refused: INTERNAL.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "x-error-codes": [
              "INTERNAL"
            ]
          }
        },
        "x-destructive-gate": false,
        "x-idempotency-key": "none",
        "x-optimistic-concurrency": null,
        "x-csrf-required": true,
        "x-authenticated": true
      }
    },
...
  },
...
}